Control.Lens.Aeson

Description

Documentation

_Object :: Prism' Value (HashMap Text Value)

_Array :: Prism' Value (Vector Value)

class AsNumber t where

Methods

_Number :: Prism' t Number

_Double :: Prism' t Double

Prism into an Double over a Value, Primitive or Number

_Integer :: Prism' t Integer

Prism into an Integer over a Value, Primitive or Number

integralValue :: (AsNumber t, Integral a) => Prism' t a

Access Integer Values as Integrals.

defined as `integer . integral'

nonNull :: Prism' Value Value

Prism into non-Null values

class AsNumber t => AsPrimitive t where

Methods

_Primitive :: Prism' t Primitive

_String :: Prism' t Text

_Bool :: Prism' t Bool

_Null :: Prism' t ()

nth :: Int -> IndexedTraversal' Int Value Value

Like ix, but for Arrays with Int indexes

decoded :: (FromJSON a, ToJSON a) => Prism' ByteString a