Control.Lens.Aeson

Description

Numbers

class AsNumber t where

Methods

_Number :: Prism' t Number

_Double :: Prism' t Double

Prism into an Double over a Value, Primitive or Number

_Integer :: Prism' t Integer

Prism into an Integer over a Value, Primitive or Number

integralValue :: (AsNumber t, Integral a) => Prism' t a

Access Integer Values as Integrals.

defined as `integer . integral'

nonNull :: Prism' Value Value

Prism into non-Null values

Primitive

data Primitive

Primitives of Value

Constructors

StringPrim !Text

NumberPrim !Number

BoolPrim !Bool

NullPrim

class AsNumber t => AsPrimitive t where

Methods

_Primitive :: Prism' t Primitive

_String :: Prism' t Text

_Bool :: Prism' t Bool

_Null :: Prism' t ()

Objects and Arrays

class AsPrimitive t => AsValue t where

Methods

_Value :: Prism' t Value

>>> putStrLn $ "{\"a\": 1, \"b\": 3}" & key "a"._Integer *~ 100
{"a":100,"b":3}

_Object :: Prism' t (HashMap Text Value)

_Array :: Prism' t (Vector Value)

key :: AsValue t => Text -> IndexedTraversal' Text t Value

Like ix, but for Object with Text indices. This often has better inference than ix when used with OverloadedStrings

nth :: AsValue t => Int -> IndexedTraversal' Int t Value

Like ix, but for Arrays with Int indexes

Decoding

class AsJSON t where

Methods

_JSON :: (FromJSON a, ToJSON a) => Prism' t a

A Prism into Value on lazy ByteStrings.